test: separate tests
